pyproject.toml
[tool.poetry]
name = "trash-calendar"
version = "0.1.0"
description = "Google Calendar integration to help keeping track of trash days."
authors = ["Kamil Nienałtowski"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.0"
httpx = "^0.25.1"
oauthlib = "^3.2.2"
pyopenssl = "^23.3.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
bs4 = "^0.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
